rename jsonschema_flatter file
